refactor: recoil persist 관련 처리 파일 디렉터리 이동 반영
